Make MobileHeader an external component
