+++ src/pages/Mobile.jsx
@@ -1,12 +1,4 @@
 import React from 'react';
-import {
-    Grid,
-    Typography,
-    makeStyles,
-    AppBar,
-    Select,
-    MenuItem,
-} from '@material-ui/core';
 
 import { Route, Switch, Redirect } from 'react-router-dom';
 
@@ -18,46 +10,8 @@ import { Contact } from '../pages/Contact';
 import { Resume } from '../pages/Resume';
 import { Projects } from '../pages/Projects';
 
+import { MobileHeader } from '../components/MobileHeader';
 import { Footer } from '../components/Footer';
-
-import { useHistory } from 'react-router-dom';
-
-const MobileHeader = props => {
-    let history = useHistory();
-    return (
-        <AppBar
-            position="static"
-            color="transparent"
-            style={{
-                // position: 'absolute',
-                // top: '0',
-                border: '1px solid black',
-                shadows: 'none',
-                height: '4vh',
-            }}
-        >
-            <Select
-                native
-                autoWidth
-                defaultValue="Home"
-                style={{ height: '4vh' }}
-                onChange={event => {
-                    let item = event.target.value;
-                    console.log(item);
-                    if (item === 'Home') history.push('/');
-                    else if (item === 'Resumé') history.push('/resume');
-                    else history.push('/' + item.toLowerCase());
-                }}
-            >
-                {['Home', 'About', 'Projects', 'Resumé', 'Contact'].map(
-                    page => (
-                        <option>{page}</option>
-                    )
-                )}
-            </Select>
-        </AppBar>
-    );
-};
 
 export const Mobile = props => {
     return (
